refactor(ui): extract ShipmentsPanel; shipments page becomes shell glue
ShipmentsPanel.svelte owns the matrix/drafts/actions/styles previously
inlined in JobShipmentsPage.svelte, keying its own deliverable/shipment
loads off job.job_id and calling onJobChange() after mutations that
affect job-level rollups (pick-up, discard, save). JobShipmentsPage is
now glue: it fetches job+contact and renders JobShell + ShipmentsPanel,
matching the pattern later job-section tasks will copy.
